requirements.txt
requests~=2.31.0
matplotlib~=3.8.4
python-dotenv~=1.0.1